moved `resolve` back into `compile`
